+++ Code/Project_B/CodeServerSide/Enums/BrokerType.cs
@@ -2,7 +2,7 @@
 
 namespace Project_B.CodeServerSide.Enums {
     [Flags]
-    public enum BrokerType : short {
+    public enum BrokerType : int {
         Default = 0,
         RedBlue = 0x01,
         GrayBlue = 0x02,
